+++ integration_tests/src/base_node.rs
@@ -33,10 +33,8 @@ use tari_shutdown::Shutdown;
 use tokio::task;
 
 use crate::{
-    utils::{
-        helpers::{get_os_assigned_ports, wait_listener_on_local_port},
-        logging::get_base_dir_for_scenario,
-    },
+    helpers::{get_os_assigned_ports, wait_listener_on_local_port},
+    logging::get_base_dir_for_scenario,
     TariWorld,
 };
