+++ integration_tests/src/lib.rs
@@ -20,6 +20,22 @@
 //   WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE
 //   USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
 
+use std::time::{Duration, Instant};
+
+use base_node::BaseNodeProcess;
+use cucumber::gherkin::Scenario;
+use http_server::MockHttpServer;
+use indexer::IndexerProcess;
+use indexmap::IndexMap;
+use miner::MinerProcess;
+use tari_common_types::types::PublicKey;
+use tari_crypto::ristretto::{RistrettoComSig, RistrettoSecretKey};
+use tari_validator_node_cli::versioned_substate_address::VersionedSubstateAddress;
+use template::RegisteredTemplate;
+use validator_node::ValidatorNodeProcess;
+use wallet::WalletProcess;
+use wallet_daemon::DanWalletDaemonProcess;
+
 pub mod base_node;
 pub mod helpers;
 pub mod http_server;
@@ -32,3 +48,120 @@ pub mod validator_node_cli;
 pub mod wallet;
 pub mod wallet_daemon;
 pub mod wallet_daemon_cli;
+
+#[derive(Debug, Default, cucumber::World)]
+pub struct TariWorld {
+    pub base_nodes: IndexMap<String, BaseNodeProcess>,
+    pub wallets: IndexMap<String, WalletProcess>,
+    pub validator_nodes: IndexMap<String, ValidatorNodeProcess>,
+    pub indexers: IndexMap<String, IndexerProcess>,
+    pub vn_seeds: IndexMap<String, ValidatorNodeProcess>,
+    pub miners: IndexMap<String, MinerProcess>,
+    pub templates: IndexMap<String, RegisteredTemplate>,
+    pub outputs: IndexMap<String, IndexMap<String, VersionedSubstateAddress>>,
+    pub http_server: Option<MockHttpServer>,
+    pub template_mock_server_port: Option<u16>,
+    pub current_scenario_name: Option<String>,
+    pub commitments: IndexMap<String, Vec<u8>>,
+    pub commitment_ownership_proofs: IndexMap<String, RistrettoComSig>,
+    pub rangeproofs: IndexMap<String, Vec<u8>>,
+    pub addresses: IndexMap<String, String>,
+    pub num_databases_saved: usize,
+    pub account_keys: IndexMap<String, (RistrettoSecretKey, PublicKey)>,
+    pub claim_public_keys: IndexMap<String, PublicKey>,
+    pub wallet_daemons: IndexMap<String, DanWalletDaemonProcess>,
+}
+
+impl TariWorld {
+    pub fn get_mock_server(&self) -> &MockHttpServer {
+        self.http_server.as_ref().unwrap()
+    }
+
+    pub fn get_miner(&self, name: &str) -> &MinerProcess {
+        self.miners
+            .get(name)
+            .unwrap_or_else(|| panic!("Miner {} not found", name))
+    }
+
+    pub fn get_wallet(&self, name: &str) -> &WalletProcess {
+        self.wallets
+            .get(name)
+            .unwrap_or_else(|| panic!("Wallet {} not found", name))
+    }
+
+    pub fn get_base_node(&self, name: &str) -> &BaseNodeProcess {
+        self.base_nodes
+            .get(name)
+            .unwrap_or_else(|| panic!("Base node {} not found", name))
+    }
+
+    pub fn get_account_component_address(&self, name: &str) -> Option<VersionedSubstateAddress> {
+        let all_components = self
+            .outputs
+            .get(name)
+            .unwrap_or_else(|| panic!("Account component address {} not found", name));
+        all_components.get("components/Account").cloned()
+    }
+
+    pub fn after(&mut self, _scenario: &Scenario) {
+        let _drop = self.http_server.take();
+
+        for (name, mut p) in self.indexers.drain(..) {
+            println!("Shutting down indexer {}", name);
+            p.shutdown.trigger();
+        }
+
+        for (name, mut p) in self.validator_nodes.drain(..) {
+            println!("Shutting down validator node {}", name);
+            p.shutdown.trigger();
+        }
+
+        for (name, mut p) in self.wallets.drain(..) {
+            println!("Shutting down wallet {}", name);
+            p.shutdown.trigger();
+        }
+        for (name, mut p) in self.base_nodes.drain(..) {
+            println!("Shutting down base node {}", name);
+            // You have explicitly trigger the shutdown now because of the change to use Arc/Mutex in tari_shutdown
+            p.shutdown.trigger();
+        }
+        for (name, mut p) in self.wallet_daemons.drain(..) {
+            println!("Shutting down wallet daemon {}", name);
+            // You have explicitly trigger the shutdown now because of the change to use Arc/Mutex in tari_shutdown
+            p.shutdown.trigger();
+        }
+        self.outputs.clear();
+        self.miners.clear();
+    }
+
+    pub async fn wait_until_base_nodes_have_transaction_in_mempool(&self, min_tx_count: usize, timeout: Duration) {
+        let timer = Instant::now();
+        'outer: loop {
+            for bn in self.base_nodes.values() {
+                let mut client = bn.create_client();
+                let tx_count = client.get_mempool_transaction_count().await.unwrap();
+
+                if tx_count < min_tx_count {
+                    // println!(
+                    //     "Waiting for {} to have {} transaction(s) in mempool (currently has {})",
+                    //     bn.name, min_tx_count, tx_count
+                    // );
+                    if timer.elapsed() > timeout {
+                        println!(
+                            "Timed out waiting for base node {} to have {} transactions in mempool",
+                            bn.name, min_tx_count
+                        );
+                        panic!(
+                            "Timed out waiting for base node {} to have {} transactions in mempool",
+                            bn.name, min_tx_count
+                        );
+                    }
+                    tokio::time::sleep(Duration::from_secs(1)).await;
+                    continue 'outer;
+                }
+            }
+
+            break;
+        }
+    }
+}
