+++ integration_tests/src/validator_node.rs
@@ -40,10 +40,8 @@ use tari_validator_node_client::ValidatorNodeClient;
 use tokio::task;
 
 use crate::{
-    utils::{
         helpers::{check_join_handle, get_os_assigned_port, get_os_assigned_ports, wait_listener_on_local_port},
         logging::get_base_dir_for_scenario,
-    },
     TariWorld,
 };
