+++ integration_tests/src/wallet_daemon_cli.rs
@@ -44,29 +44,16 @@ use tari_transaction_manifest::{parse_manifest, ManifestValue};
 use tari_validator_node_cli::command::transaction::CliArg;
 use tari_wallet_daemon_client::{
     types::{
-        AccountGetResponse,
-        AccountsCreateFreeTestCoinsRequest,
-        AccountsCreateRequest,
-        AccountsGetBalancesRequest,
-        AuthLoginAcceptRequest,
-        AuthLoginRequest,
-        AuthLoginResponse,
-        ClaimBurnRequest,
-        ClaimBurnResponse,
-        ConfidentialTransferRequest,
-        MintAccountNFTRequest,
-        ProofsGenerateRequest,
-        RevealFundsRequest,
-        TransactionSubmitRequest,
-        TransactionWaitResultRequest,
-        TransferRequest,
+        AccountGetResponse, AccountsCreateFreeTestCoinsRequest, AccountsCreateRequest, AccountsGetBalancesRequest,
+        AuthLoginAcceptRequest, AuthLoginRequest, AuthLoginResponse, ClaimBurnRequest, ClaimBurnResponse,
+        ConfidentialTransferRequest, MintAccountNFTRequest, ProofsGenerateRequest, RevealFundsRequest,
+        TransactionSubmitRequest, TransactionWaitResultRequest, TransferRequest,
     },
-    ComponentAddressOrName,
-    WalletDaemonClient,
+    ComponentAddressOrName, WalletDaemonClient,
 };
 
 use super::wallet_daemon::get_walletd_client;
-use crate::{utils::validator_node_cli::add_substate_addresses, TariWorld};
+use crate::{validator_node_cli::add_substate_addresses, TariWorld};
 
 pub async fn claim_burn(
     world: &mut TariWorld,
@@ -766,7 +753,7 @@ pub(crate) async fn get_wallet_daemon_client(world: &TariWorld, wallet_daemon_na
     get_walletd_client(port).await
 }
 
-pub(crate) async fn get_auth_wallet_daemon_client(world: &TariWorld, wallet_daemon_name: &str) -> WalletDaemonClient {
+pub async fn get_auth_wallet_daemon_client(world: &TariWorld, wallet_daemon_name: &str) -> WalletDaemonClient {
     let mut client = get_wallet_daemon_client(world, wallet_daemon_name).await;
     // authentication
     let AuthLoginResponse { auth_token } = client
